refactor(tests): Add parametrization to smoke, shape, and gradcheck tests (#3419)
- tests/feature/test_steerer.py::test_shape: Add dtype parameter
- tests/augmentation/_3d/test_random_horizontal_flip_3d.py::test_smoke: Add device parameter
- tests/augmentation/_3d/test_random_horizontal_flip_3d.py::test_gradcheck: Enforce float64 for gradient accuracy
- tests/integration/test_warp.py::test_smoke: Add device parameter
- tests/utils/test_misc.py: Enforce float64 in 3 test_gradcheck methods
- tests/utils/test_misc.py: Add torch import for float64 reference

Scope: 5 files, 8 methods, minimal changes to enable conftest.py parametrization
No structural changes or new imports beyond torch (already used elsewhere)

+++ tests/utils/test_misc.py
@@ -15,6 +15,8 @@
 # limitations under the License.
 #
 
+import torch
+
 from kornia.core import tensor
 from kornia.utils.misc import (
     differentiable_clipping,
@@ -34,7 +36,7 @@ class TestDifferentiableClipping(BaseTester):
         self.assert_close(y, y_expected)
 
     def test_gradcheck(self, device):
-        x = tensor([1.0, 6.0, 11.0, 12.0], device=device)
+        x = tensor([1.0, 6.0, 11.0, 12.0], device=device, dtype=torch.float64)
         self.gradcheck(differentiable_clipping, (x, 5.0, 10.0))
 
 
@@ -47,7 +49,7 @@ class TestDifferentiablePolynomialRounding(BaseTester):
         self.assert_close(y, y_expected)
 
     def test_gradcheck(self, device):
-        x = tensor([1.0, 6.0, 10.0, 12.0], device=device)
+        x = tensor([1.0, 6.0, 10.0, 12.0], device=device, dtype=torch.float64)
         self.gradcheck(differentiable_polynomial_rounding, (x))
 
 
@@ -60,5 +62,5 @@ class TestDifferentiablePolynomialFloor(BaseTester):
         self.assert_close(y, y_expected)
 
     def test_gradcheck(self, device):
-        x = tensor([1.5, 3.1, 5.9, 6.6], device=device)
+        x = tensor([1.5, 3.1, 5.9, 6.6], device=device, dtype=torch.float64)
         self.gradcheck(differentiable_polynomial_floor, (x))
